// lib/interfaces.tsx
interface motherboard {
  id: string;
  price: number | null;
  socket: string | null;
  form_factor: string | null;
  max_memory: number | null;
  memory_slots: number | null;
  color: string | null;
  createdAt: string | null;
  updatedAt: string | null;
  manufacturer: string | null;
  name: string | null;
}